# Faster time coordinate categorization