## 🚀 Pull Request

### Description
<!-- Provide a clear description about your awesome pull request -->
<!-- Tell us all about your new feature, improvement, or bug fix -->

Make time coordinate categorization faster by converting the time points to datetimes in a single call.

Example script:
```python
In [1]: import numpy as np

In [2]: import cf_units

In [3]: from iris.cube import Cube

In [4]: from iris.coords import DimCoord

In [5]: from iris.coord_categorisation import add_year

In [6]: coord = DimCoord(np.arange(10000), standard_name='time', units=cf_units.Unit("days since epoch", "standard"))

In [7]: cube = Cube(np.empty(10000), dim_coords_and_dims=[(coord, 0)])

In [8]: %timeit add_year(cube.copy(), 'time')
9.99 ms ± 24.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

while with the current `main` branch, the timing result is
```python
345 ms ± 2.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

---
[Consult Iris pull request check list]( https://scitools-iris.readthedocs.io/en/latest/developers_guide/contributing_pull_request_checklist.html)

---
Add any of the below labels to trigger actions on this PR:

- https://github.com/SciTools/iris/labels/benchmark_this