@bouweandela 2024-06-11
Would it make sense to add the example code above as a benchmark?